change "c:geo" map label back to "map", fix multiple caches being shown when selecting a single cache in the cache list context menu
fixes #205

+++ src/cgeo/geocaching/apps/cachelist/CacheListApp.java
@@ -12,6 +12,6 @@ interface CacheListApp extends App {
 
 	boolean invoke(final cgGeo geo, final List<cgCache> caches,
 			final Activity activity, final Resources res,
-			final Integer searchId);
+			final Long searchId);
 
 }
